add write_as property to boolean elements
SSh model is quite old and did not use the 'write_as' property now
provided by  Config::Model::Value.

This property allows a boolean value to be read or written as something
else than 0 and 1. In Ssh case, the boolean value must be written as
yes/no. This conversion was done by adhoc code in the backend.

When setting the boolean value with "write_as => ['no','yes']", this
conversion is not longer necessary and was removed from backend code.

The tests are also updated according to this change